It finally works ))

+++ src/main/java/ru/kata/spring/boot_security/demo/repositories/UserRepositoryImp.java
@@ -41,8 +41,7 @@ public class UserRepositoryImp implements UserRepository{
 
     @Override
     public void deleteUserById(Long id) {
-        manager.createQuery ("delete from User " +
-                "where id=:userID")
+        manager.createQuery ("delete from User " + "where id=:userID")
                 .setParameter("userID", id).executeUpdate();
     }
 }
